roles/vx/files/mqtt2out/requirements.txt
pyyaml
paho-mqtt
influxdb


